Insures installation failure if dtplyr is not installed.
I also moved around the `library` calls in the `global.R` file in an attempt
to better protect against polutting the users workspace with functions from
packages that `multiGSEA.shiny` depends on.

These are guarded by an `interactive()` test, which I assume will be `FALSE`
when running as a solo shiny when the same app is deployed on a shiny server.

Fixes #2, thanks to @tomsing1 for reporting.